+++ src/main/java/org/iq47/model/OrderAddressRepository.java
@@ -4,6 +4,6 @@ import org.iq47.model.entity.OrderAddress;
 import org.iq47.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderAddressRepository extends JpaRepository<OrderAddress, Long> {
+public interface OrderAddressRepository extends JpaRepository<OrderAddress, Integer> {
     OrderAddress getByUser(User user);
 }
