+++ src/main/java/org/iq47/network/request/ShopItemCreateRequest.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class ShopItemCreateRequest {
     private String itemName;
-    private long shop_id;
+    private int shop_id;
     private int price;
 
 }
